Added IShapeAreaService interface

+++ GeomertyShapeLib/Models/Circle.cs
@@ -2,9 +2,9 @@
 
 public sealed class Circle : ShapeBase
 {
-    public int Radius { get; set; }
+    public double Radius { get; set; }
 
-    public Circle(int radius)
+    public Circle(double radius)
     {
         Radius = radius;
     }
